requirements.txt
flask
requests
anthropic
python-dotenv
pypdf
fpdf2
gunicorn
